CrystalSpider/requirements.txt
Bio==1.4.0
expertai-nlapi==2.4.1
fastapi==0.86.0
grobid_client==0.8.2
gunicorn==20.1.0
pymongo==4.3.2
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
requests_html==0.10.0
uvicorn==0.19.0
Office365-REST-Python-Client==2.4.0
httpx~=0.22.0
beautifulsoup4~=4.11.1
pypeln~=0.4.9
wheel~=0.37.1
bs4~=0.0.1
loguru~=0.6.0
meilisearch==0.22.2
stop-words
python-multipart==0.0.5
itsdangerous==2.1.2
dnspython
llama-index==0.5.3
langchain~=0.0.126
fastapi-sessions
starlette~=0.20.4
fastapi-events
scipy~=1.10.1
numpy~=1.24.2
faiss-cpu
retrying~=1.3.4
Jinja2==3.1.2